+++ processor/server.py
@@ -4,8 +4,9 @@ import mediacloud.error
 from flask import render_template, jsonify
 import json
 from typing import Dict, List
+from itertools import chain
 
-from processor import create_flask_app, VERSION, get_mc_client
+from processor import create_flask_app, VERSION, get_mc_client, PLATFORMS, SOURCE_MEDIA_CLOUD
 from processor.projects import load_project_list
 import processor.database.stories_db as stories_db
 
@@ -23,7 +24,10 @@ app.jinja_env.filters['as_pretty_json'] = as_pretty_json
 @app.route("/", methods=['GET'])
 def home():
     projects = load_project_list()
-    return render_template('home.html', projects=projects, version=VERSION)
+    # show overall ingest over last two weeks
+    data_for_graph = _prep_for_graph([stories_db.stories_by_published_day(platform=p, limit=30) for p in PLATFORMS],
+                                     PLATFORMS)
+    return render_template('home.html', projects=projects, version=VERSION, ingest_data=data_for_graph)
 
 
 @app.route("/update-config", methods=['POST'])
@@ -32,22 +36,17 @@ def update_config():
     return jsonify(config)
 
 
-def _prep_for_graph(counts_1: List, count_2: List, type1: str, type2: str) -> List[Dict]:
-    data1 = {r['day'].strftime("%Y-%m-%d"): r['stories'] for r in counts_1}
-    data2 = {r['day'].strftime("%Y-%m-%d"): r['stories'] for r in count_2}
-    dates = set(data1.keys() | data2.keys())
+def _prep_for_graph(counts: List[List], names: List[str]) -> List[Dict]:
+    cleaned_data = [{r['day'].strftime("%Y-%m-%d"): r['stories'] for r in series} for series in counts]
+    dates = set(chain(*[series.keys() for series in cleaned_data]))
     stories_by_day_data = []
     for d in dates:  # need to make sure there is a pair of entries for each date
-        stories_by_day_data.append(dict(
-            date=d,
-            type=type1,
-            count=data1[d] if d in data1 else 0
-        ))
-        stories_by_day_data.append(dict(
-            date=d,
-            type=type2,
-            count=data2[d] if d in data2 else 0
-        ))
+        for idx, series in enumerate(cleaned_data):
+            stories_by_day_data.append(dict(
+                date=d,
+                type=names[idx],
+                count=series[d] if d in series else 0
+            ))
     return stories_by_day_data
 
 
@@ -55,9 +54,9 @@ def _prep_for_graph(counts_1: List, count_2: List, type1: str, type2: str) -> Li
 def project_processed_by_day(project_id_str):
     project_id = int(project_id_str)
     data = _prep_for_graph(
-        stories_db.stories_by_processed_day(project_id, True, None),
-        stories_db.stories_by_processed_day(project_id, False, None),
-        'above threshold', 'below threshold'
+        [stories_db.stories_by_processed_day(project_id, True, None),
+         stories_db.stories_by_processed_day(project_id, False, None)],
+        ['above threshold', 'below threshold']
     )
     return jsonify(data)
 
@@ -66,9 +65,9 @@ def project_processed_by_day(project_id_str):
 def project_published_by_day(project_id_str):
     project_id = int(project_id_str)
     data = _prep_for_graph(
-        stories_db.stories_by_published_day(project_id, True),
-        stories_db.stories_by_published_day(project_id, False),
-        'above threshold', 'below threshold'
+        [stories_db.stories_by_published_day(project_id=project_id, above_threshold=True),
+         stories_db.stories_by_published_day(project_id=project_id, above_threshold=False)],
+        ['above threshold', 'below threshold']
     )
     return jsonify(data)
 
@@ -77,9 +76,9 @@ def project_published_by_day(project_id_str):
 def project_posted_by_day(project_id_str):
     project_id = int(project_id_str)
     data = _prep_for_graph(
-        stories_db.stories_by_processed_day(project_id, True, True),
-        stories_db.stories_by_processed_day(project_id, True, False),
-        'sent to main server', 'not sent to main server'
+        [stories_db.stories_by_processed_day(project_id, True, True),
+         stories_db.stories_by_processed_day(project_id, True, False)],
+        ['sent to main server', 'not sent to main server']
     )
     return jsonify(data)
 
@@ -90,26 +89,25 @@ def a_project(project_id_str):
     # pull out the project info
     project = [p for p in load_project_list() if p['id'] == project_id][0]
 
-    # get some stats
+    # show overall ingest over last two weeks
+    data_for_graph = _prep_for_graph([stories_db.stories_by_published_day(platform=p, project_id=project_id, limit=30)
+                                      for p in PLATFORMS], PLATFORMS)
 
     # show some recent story results
+    stories_above = stories_db.recent_stories(project_id, True)
+    stories_below = stories_db.recent_stories(project_id, False)
+    # story_ids = list(set([s.stories_id for s in stories_above]) | set([s.stories_id for s in stories_below]))
     try:
-        stories_above = stories_db.recent_stories(project_id, True)
-        stories_below = stories_db.recent_stories(project_id, False)
-        story_ids = list(set([s.stories_id for s in stories_above]) | set([s.stories_id for s in stories_below]))
+        mc_story_ids = list(set([s.stories_id for s in stories_above if s.source == SOURCE_MEDIA_CLOUD])
+                            | set([s.stories_id for s in stories_below if s.source == SOURCE_MEDIA_CLOUD]))
         mc = get_mc_client()
-        if len(story_ids) > 0:
-            stories = mc.storyList("stories_id:({})".format(" ".join([str(s) for s in story_ids])))
+        if len(mc_story_ids) > 0:
+            stories = mc.storyList("stories_id:({})".format(" ".join([str(s) for s in mc_story_ids])))
             story_lookup = {s['stories_id']: s for s in stories}
         else:
             story_lookup = {}
-        # sometimes Media Cloud doesn't return info for a story :-(
-        stories_above = [s for s in stories_above if s.stories_id in story_lookup]
-        stories_below = [s for s in stories_below if s.stories_id in story_lookup]
     except mediacloud.error.MCException as mce:
         # we currently can't query for BIGINT story ids with a stories_id clause in the query
-        stories_above = []
-        stories_below = []
         story_lookup = {}
         pass
 
@@ -124,6 +122,7 @@ def a_project(project_id_str):
 
     # render it all
     return render_template('project.html',
+                           ingest_data=data_for_graph,
                            unposted_above_story_count=unposted_above_story_count,
                            above_threshold_pct=above_threshold_pct,
                            posted_above_story_count=posted_above_story_count,